backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pyjwt==2.8.0
werkzeug==2.3.0
sqlalchemy==1.4.46
flask-sqlalchemy==2.5.1
